requirements.txt
aiohttp-socks==0.8.4; python_full_version >= "3.7.1" and python_version < "4.0"
fastapi==0.103.2; python_full_version >= "3.7.1" and python_version < "4.0"
requests==2.31.0; python_full_version >= "3.7.1" and python_version < "4.0"
uvicorn==0.22.0; python_full_version >= "3.7.1" and python_version < "4.0"
typer==0.9.0; python_full_version >= "3.7.1" and python_version < "4.0"
click==8.1.7; python_full_version >= "3.7.1" and python_version < "4.0"
PyYAML==6.0.1; python_full_version >= "3.7.1" and python_version < "4.0"
pydantic==1.8.2; python_full_version >= "3.7.1" and python_version < "4.0"